remove overlay controller
